Bug 1004098 - Make nsTArray use size_t in its interface (32bitness is fine as an internal detail) - r=froydnj, sr=bsmedberg
UltraBlame original commit: b347f6eb2239cf0c08eef7af4f6d9eb36a0eda99

+++ content/media/AudioNodeExternalInputStream.h
@@ -92,8 +92,8 @@ private:
 
 
 
-  uint32_t GetTrackMapEntry(const StreamBuffer::Track& aTrack,
-                            GraphTime aFrom);
+  size_t GetTrackMapEntry(const StreamBuffer::Track& aTrack,
+                          GraphTime aFrom);
 };
 
 }
