+++ content/media/webaudio/WebAudioUtils.cpp
@@ -16,7 +16,7 @@ namespace dom {
 
 
 
-const uint32_t WebAudioUtils::MaxChannelCount = 32;
+const size_t WebAudioUtils::MaxChannelCount = 32;
 
 struct ConvertTimeToTickHelper
 {
