+++ image/src/RasterImage.cpp
@@ -2861,7 +2861,7 @@ RasterImage::RequestDiscard()
 
 
 nsresult
-RasterImage::DecodeSomeData(uint32_t aMaxBytes, DecodeStrategy aStrategy)
+RasterImage::DecodeSomeData(size_t aMaxBytes, DecodeStrategy aStrategy)
 {
   
   NS_ABORT_IF_FALSE(mDecoder, "trying to decode without decoder!");
@@ -2886,8 +2886,8 @@ RasterImage::DecodeSomeData(uint32_t aMaxBytes, DecodeStrategy aStrategy)
   MOZ_ASSERT(mBytesDecoded < mSourceData.Length());
 
   
-  uint32_t bytesToDecode = std::min(aMaxBytes,
-                                    mSourceData.Length() - mBytesDecoded);
+  size_t bytesToDecode = std::min(aMaxBytes,
+                                  mSourceData.Length() - mBytesDecoded);
   nsresult rv = WriteToDecoder(mSourceData.Elements() + mBytesDecoded,
                                bytesToDecode,
                                aStrategy);
@@ -3384,7 +3384,7 @@ RasterImage::DecodePool::DecodeJob::Run()
 
   mRequest->mRequestStatus = DecodeRequest::REQUEST_ACTIVE;
 
-  uint32_t oldByteCount = mImage->mBytesDecoded;
+  size_t oldByteCount = mImage->mBytesDecoded;
 
   DecodeType type = DECODE_TYPE_UNTIL_DONE_BYTES;
 
@@ -3396,7 +3396,7 @@ RasterImage::DecodePool::DecodeJob::Run()
 
   DecodePool::Singleton()->DecodeSomeOfImage(mImage, DECODE_ASYNC, type, mRequest->mBytesToDecode);
 
-  uint32_t bytesDecoded = mImage->mBytesDecoded - oldByteCount;
+  size_t bytesDecoded = mImage->mBytesDecoded - oldByteCount;
 
   mRequest->mRequestStatus = DecodeRequest::REQUEST_WORK_DONE;
