+++ image/src/RasterImage.h
@@ -369,7 +369,7 @@ private:
 
     RasterImage* mImage;
 
-    uint32_t mBytesToDecode;
+    size_t mBytesToDecode;
 
     enum DecodeRequestStatus
     {
@@ -681,7 +681,7 @@ private:
   
   nsRefPtr<Decoder>          mDecoder;
   nsRefPtr<DecodeRequest>    mDecodeRequest;
-  uint32_t                   mBytesDecoded;
+  size_t                     mBytesDecoded;
 
   bool                       mInDecoder;
   
@@ -729,7 +729,7 @@ private:
   nsresult SyncDecode();
   nsresult InitDecoder(bool aDoSizeDecode);
   nsresult WriteToDecoder(const char *aBuffer, uint32_t aCount, DecodeStrategy aStrategy);
-  nsresult DecodeSomeData(uint32_t aMaxBytes, DecodeStrategy aStrategy);
+  nsresult DecodeSomeData(size_t aMaxBytes, DecodeStrategy aStrategy);
   bool     IsDecodeFinished();
   TimeStamp mDrawStartTime;
